Added ScavengerHandler
ScavengerHandler will bridge the gap between the TwiMLHandler and our
code.  Ideally we should be able to isolate this so that people could
interact with the system using only a data plan or just by email or
whatever else.  I guess - ScavengerHandler should return responses and
not care about the format so much as the content.